backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "brotli-asgi>=1.4.0",
    "fast-langdetect>=0.3.1",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.115.11",
    "pydantic>=2.10.6",
    "pydantic-to-typescript>=2.0.0",
    "zstd>=1.5.6.6",
]
